Addresses #159. Redirecting subdomain SSL requests to non-SSL.
Attempting to access any bigbread.net subdomain via https lands the
user on the production site without any indication that this is the
case. Since this seems awfully dangerous, we're redirecting to the
non-SSL host which puts us where we want, though without SSL. Doing
this at the .htaccess level means that production has different
content than subdomains, so the deploy scripts were updated as well.