Fixing a bug in the contractor registration process.
When going back to the info page, making a change and hitting next, the
user id value wasn't being set so the system attempted to do an insert.
Also, the password validation was complaining. When editing, you can now
leave the password value blank and no change will get made.