Changing how addresses are associated.
Making more important models belongsTo Address gets pretty unwieldy when
you want to delete the important model, but the Address record gets in
the way. Instead, I made the Address model polymorphic so that a
Building, for example, can hasOne Address. This means that I can set the
association to be dependent and deleting the building will automagically
delete the associated address. No more garbage.