Handling addresses at the model level.
Now that the Address model has gone polymorphic, we need to be sure that
we save the model name. In both cases so far (Building, Contractor),
address data is being saved via a call to Model::saveAll(). Rather than
changing the data at the point of entry, I've just overridden
Model::saveAll() to insert the proper model alias in the model field.