Anonymous API call no longer remains logged in.
Every API request must be authenticated. Since we use our own API for our
internal ajax requests, the system was authenticating if there was no existing
session. This meant that API requests from public pages (like the zip code
changer on the homepage) were being authenticated. The system was not, however,
logging the user out at the end of the request.