Setting a location context in the session.
In order to preserve state when hopping between ways to save and the dashboard
where explicit locations are optional, we're setting the last accessed location
so that it will be used in those cases where no location is explicitly
requested. If no context exists either, then we'll fallback to what we were
doing before.